fix: add spec-helper to its own file for demo script
